requirements.txt
Flask>=3.0.0
matplotlib>=3.9.0
numpy>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.0
